Rework handling and listen with goroutines
Spawn a new goroutine for each listening FIFO and use a
standard blocking read rather than using polled non-blocking
I/O. This is simpler and should be far more efficient.

Use switch statements, not if/else if chains.

Since the Msg struct contains all of the necessary data for
dispatch, we make handleMsg() into a method of Msg, Send().